+++ R/tt_pos_and_access.R
@@ -295,25 +295,43 @@ setMethod(
     if (obj_name(tt) == path[1]) {
       path <- path[-1]
     }
-    cur <- list(tt)
-    curpath <- path
-    while (length(curpath > 0)) {
-      for (n_id_levs in seq_along(cur)) {
-        kids <- tree_children(cur[[n_id_levs]])
-        curname <- curpath[1]
-        if (curname == "@content") {
-          cur <- content_table(cur)
-        } else if (curname %in% names(kids)) {
-          cur <- kids[names(kids) == curname]
-        } else if (n_id_levs == length(cur)) {
-          stop("Path appears invalid for this tree at step ", curname)
-        }
-      }
-      curpath <- curpath[-1]
-    }
-    cur
+    # cur_tbl <- tt
+    # cur_path <- path
+    .extract_through_path(tt, path)
   }
 )
+
+.extract_through_path <- function(cur_tbl, cur_path, no_stop = FALSE) {
+  while (length(cur_path > 0)) {
+    kids <- tree_children(cur_tbl)
+    curname <- cur_path[1]
+    if (curname == "@content") {
+      cur_tbl <- content_table(cur_tbl)
+    } else if (curname %in% names(kids)) {
+      cur_tbl <- kids[names(kids) == curname]
+      if (length(cur_tbl) > 1 && length(cur_path) > 1) {
+        # cur_tbl_tmp <- cur_tbl
+        # cur_tbl <- cur_tbl[[1]]
+        # cur_path <- cur_path[-1]
+        cur_tbl <- sapply(cur_tbl, function(cti) .extract_through_path(cti, cur_path[-1], no_stop = TRUE))
+        found_values <- !sapply(cur_tbl, is.null)
+        cur_tbl <- cur_tbl[found_values]
+        if (sum(found_values) == 1) {
+          cur_tbl <- cur_tbl[[1]]
+        }
+        cur_path <- cur_path[1]
+      } else {
+        cur_tbl <- cur_tbl[[1]]
+      }
+    } else if (!no_stop) {
+      stop("Path appears invalid for this tree at step ", curname)
+    } else {
+      return(NULL)
+    }
+    cur_path <- cur_path[-1]
+  }
+  cur_tbl
+}
 
 #' @note Setting `NULL` at a defined path removes the corresponding sub-table.
 #'